Changes to vertex math. Added benchmark test to terminal.

+++ src/Tile.java
@@ -1,13 +1,13 @@
 import java.util.Arrays;
 public class Tile {
-    //remove
-
-    //keep
     double xOffset, yOffset, tileWidth, tileLength;
     int[] visibleIndices;
     int indicesCount;
+
+
+    //remove
     int [][] localColorBuffer;
-    double [][] localDepthBuffer;
+    float [][] localDepthBuffer;
 
     Tile(double x, double y, double w, double l, int count){
         xOffset = x;
@@ -16,8 +16,10 @@ public class Tile {
         tileLength = l;
         indicesCount = 0;
         visibleIndices = new int[count];
+
         localColorBuffer = new int[(int)tileWidth][(int)tileLength];
-        localDepthBuffer = new double[(int)tileWidth][(int)tileLength];
+        localDepthBuffer = new float[(int)tileWidth][(int)tileLength];
+
         for (int i = 0; i < (int)tileWidth; i++) {
             java.util.Arrays.fill(localColorBuffer[i], 0); 
             java.util.Arrays.fill(localDepthBuffer[i], 0); 
@@ -25,6 +27,7 @@ public class Tile {
     }
     void emptyTileData(){
         indicesCount = 0;
+
         for (int i = 0; i < (int)tileWidth; i++) {
             java.util.Arrays.fill(localColorBuffer[i], 0); 
             java.util.Arrays.fill(localDepthBuffer[i], 0);
